+++ gakumas-tools/components/Simulator/Simulator.js
@@ -155,9 +155,11 @@ export default function Simulator() {
   return (
     <div id="simulator_loadout" className={styles.loadoutEditor}>
       <div className={styles.configurator}>
-        <>
-          ※ Season 25 score multiplier is inaccurate for parameters above 2400.
-        </>
+        <div>
+          ※ Season 26 preview is based only on the announced p-items and
+          criteria. Turn types are not yet known.
+        </div>
+        <div>※ Score multiplier is inaccurate for parameters above 2400.</div>
         <StageSelect />
         {stage.type == "event" ? (
           t("enterPercents")